+++ cairo1-run/src/cairo_run.rs
@@ -92,7 +92,7 @@ pub struct Cairo1RunConfig<'a> {
     /// Should be true if either air_public_input or cairo_pie_output are needed
     /// Sets builtins stop_ptr by calling `final_stack` on each builtin
     pub finalize_builtins: bool,
-    /// Appends return values to the output segment. This is performed by default when running in proof_mode
+    /// Appends the return and input values to the output segment. This is performed by default when running in proof_mode
     pub append_return_values: bool,
 }
 
@@ -111,10 +111,18 @@ impl Default for Cairo1RunConfig<'_> {
     }
 }
 
-// Runs a Cairo 1 program
-// Returns the runner after execution + the return values + the serialized return values (if serialize_output is enabled)
-// The return values will contain the memory values just as they appear in the VM, after removing the PanicResult enum (if present).
-// Except if either the flag append_return_values or proof_mode are enabled, in which case the return values will consist of its serialized form: [array_len, array[0], array[1], ..., array[array_len -1]]
+impl Cairo1RunConfig<'_> {
+    // Returns true if the flags in the config enable adding the output builtin and
+    // copying input and output values into it's segment
+    fn copy_to_output(&self) -> bool {
+        self.append_return_values || self.proof_mode
+    }
+}
+
+/// Runs a Cairo 1 program
+/// Returns the runner after execution + the return values + the serialized return values (if serialize_output is enabled)
+/// The return values will contain the memory values just as they appear in the VM, after removing the PanicResult enum (if present).
+/// Except if either the flag append_return_values or proof_mode are enabled, in which case the return values will consist of its serialized form: [array_len, array[0], array[1], ..., array[array_len -1]]
 pub fn cairo_run_program(
     sierra_program: &SierraProgram,
     cairo_run_config: Cairo1RunConfig,
@@ -147,7 +155,7 @@ pub fn cairo_run_program(
         _ => None,
     };
 
-    if (cairo_run_config.proof_mode || cairo_run_config.append_return_values)
+    if cairo_run_config.copy_to_output()
         && !check_only_array_felt_input_type(
             &main_func.signature.param_types,
             &sierra_program_registry,
@@ -155,7 +163,7 @@ pub fn cairo_run_program(
     {
         return Err(Error::IlegalInputValue);
     };
-    if (cairo_run_config.proof_mode || cairo_run_config.append_return_values)
+    if cairo_run_config.copy_to_output()
         && !check_only_array_felt_return_type(return_type_id, &sierra_program_registry)
     {
         return Err(Error::IlegalReturnValue);
@@ -193,7 +201,7 @@ pub fn cairo_run_program(
     let mut hint_processor = Cairo1HintProcessor::new(
         &processor_hints,
         RunResources::default(),
-        cairo_run_config.append_return_values || cairo_run_config.proof_mode,
+        cairo_run_config.copy_to_output(),
     );
 
     let data: Vec<MaybeRelocatable> = instructions
@@ -256,8 +264,6 @@ pub fn cairo_run_program(
 
     runner.end_run(false, false, &mut hint_processor)?;
 
-    let skip_output = cairo_run_config.proof_mode || cairo_run_config.append_return_values;
-
     let result_inner_type_size =
         result_inner_type_size(return_type_id, &sierra_program_registry, &type_sizes);
     // Fetch return values
@@ -266,11 +272,11 @@ pub fn cairo_run_program(
         result_inner_type_size,
         &runner.vm,
         builtin_count,
-        skip_output,
+        cairo_run_config.copy_to_output(),
     )?;
 
     let serialized_output = if cairo_run_config.serialize_output {
-        if cairo_run_config.append_return_values || cairo_run_config.proof_mode {
+        if cairo_run_config.copy_to_output() {
             // The return value is already serialized, so we can just print the array values
             let mut output_string = String::from("[");
             // Skip array_len
@@ -295,7 +301,7 @@ pub fn cairo_run_program(
 
     // Set stop pointers for builtins so we can obtain the air public input
     if cairo_run_config.finalize_builtins {
-        if skip_output {
+        if cairo_run_config.copy_to_output() {
             // Set stop pointer for each builtin
             runner.vm.builtins_final_stack_from_stack_pointer_dict(
                 &builtins
@@ -456,7 +462,6 @@ fn load_arguments(
         .param_types
         .iter()
         .any(|ty| ty.debug_name.as_ref().is_some_and(|n| n == "SegmentArena"));
-    let append_output = cairo_run_config.append_return_values || cairo_run_config.proof_mode;
     // This AP correction represents the memory slots taken up by the values created by `create_entry_code`:
     // These include:
     // * The builtin bases (not including output)
@@ -466,7 +471,7 @@ fn load_arguments(
     //  * info_segment_ptr
     //  * 0
     let mut ap_offset = runner.get_program().builtins_len();
-    if append_output {
+    if cairo_run_config.copy_to_output() {
         ap_offset += runner.get_program().builtins_len() - 1;
     }
     if got_segment_arena {
@@ -514,7 +519,7 @@ fn create_entry_code(
     initial_gas: usize,
     config: &Cairo1RunConfig,
 ) -> Result<(CasmContext, Vec<BuiltinName>), Error> {
-    let copy_to_output_builtin = config.proof_mode || config.append_return_values;
+    let copy_to_output_builtin = config.copy_to_output();
     let signature = &func.signature;
     // The builtins in the formatting expected by the runner.
     let (builtins, builtin_offset) =